move hooks and deprecate more
