+++ src/app/invoices/components/deprecated/invoice-document.tsx
@@ -2,9 +2,9 @@ import { forwardRef } from 'react';
 
 import { formatCurrency } from '@/lib/currency';
 
-import { Invoice } from '../v2/schema';
-import { calculateSubtotalTaxAndTotal, formatDate } from './dashboard/utils';
-import { InvoiceFormData } from './deprecated/create/invoice-form-context';
+import { Invoice } from '../../v2/schema';
+import { calculateSubtotalTaxAndTotal, formatDate } from '../dashboard/utils';
+import { InvoiceFormData } from './create/invoice-form-context';
 /**
  * Render a PDF-like document displaying the invoice details.
  *
