+++ src/app/invoices/hooks/use-invoice-form.ts
@@ -1,6 +1,6 @@
 import { useFormContext } from 'react-hook-form';
 
-import { InvoiceForm } from '../schema';
+import { InvoiceForm } from '../v2/schema';
 
 /** Hook for accessing the invoice form context */
 export const useInvoiceForm = () => useFormContext<InvoiceForm>();
