+++ src/app/invoices/v2/invoice-controls/new-invoice-tab.tsx
@@ -21,8 +21,8 @@ import { Textarea } from '@/components/ui/textarea';
 import { cn } from '@/lib/utils';
 
 import { ClientCard } from '../client-card/client-card';
-import { useClients } from '../hooks/use-clients';
-import { useInvoiceForm } from '../hooks/use-invoice-form';
+import { useClients } from '../../hooks/use-clients';
+import { useInvoiceForm } from '../../hooks/use-invoice-form';
 import { isInTheFuture } from '../schema';
 import { ClientSheet } from './client-sheet';
 import { ItemsCard } from './items-card';
